+++ src/realmrelay/game/messaging/outgoing/SetCondition.java
@@ -8,7 +8,7 @@ import java.util.function.Consumer;
 public class SetCondition extends OutgoingMessage {
 
 	public int conditionEffect;
-	public float conditionDuration;
+	public double conditionDuration;
 
 	public SetCondition(int id, Consumer callback) {
 		super(id, callback);
